removed wx_tutorial.py and divided code in multiple files
