+++ .vs/PythonSettings.json
@@ -1,0 +1,3 @@
+{
+  "Interpreter": "Global|PythonCore|3.7-32"
+}
