+++ camera.py
@@ -1,0 +1,130 @@
+from OpenGL.GL import *
+from OpenGL.GLU import *
+import numpy as np
+import random
+import webcolors
+from enums import Axis
+
+class Camera():
+    def __init__(self, x, y, z, b, c):
+        self.x = float(x)
+        self.y = float(y)
+        self.z = float(z)
+        self.b = float(b)
+        self.c = float(c)
+        self.start = (self.x, self.y, self.z, self.b, self.c)
+
+        self.angle = 0
+        self.rotationVector = []
+    
+        colors = webcolors.css3_hex_to_names
+        chosenColor = random.choice(list(colors.keys()))
+        chosenRgb = webcolors.hex_to_rgb(chosenColor)
+        self.red = chosenRgb.red
+        self.green = chosenRgb.green
+        self.blue = chosenRgb.blue
+
+        self.mode = 'normal'
+
+    def onDraw(self):
+        glPushMatrix()
+        glTranslatef(self.x, self.y, self.z)
+        if self.mode == 'normal':
+            if self.b != 0.0:
+                glRotatef(self.b, 0, 1, 0)
+            if self.c != 0.0:
+                glRotatef(self.c, 0, 0, 1)
+        elif self.mode == 'rotate':
+            glRotatef(self.angle, self.rotationVector[0], self.rotationVector[1], self.rotationVector[2])
+
+        glBegin(GL_QUADS)
+        ## bottom
+        glColor3ub(self.red, self.green, self.blue)
+        glVertex3f(-0.025, -0.05, -0.05)
+        glVertex3f( 0.025, -0.05, -0.05)
+        glVertex3f( 0.025, -0.05,  0.05)
+        glVertex3f(-0.025, -0.05,  0.05)
+
+        ## right
+        glColor3ub(self.red, self.green, self.blue)
+        glVertex3f(-0.025,  0.05, -0.05)
+        glVertex3f( 0.025,  0.05, -0.05)
+        glVertex3f( 0.025, -0.05, -0.05)
+        glVertex3f(-0.025, -0.05, -0.05)
+        
+        ## top
+        glColor3ub(self.red, self.green, self.blue)
+        glVertex3f(-0.025,  0.05, -0.05)
+        glVertex3f( 0.025,  0.05, -0.05)
+        glVertex3f( 0.025,  0.05,  0.05)
+        glVertex3f(-0.025,  0.05,  0.05)
+
+        ## left
+        glColor3ub(self.red, self.green, self.blue)
+        glVertex3f(-0.025, -0.05,  0.05)
+        glVertex3f( 0.025, -0.05,  0.05)
+        glVertex3f( 0.025,  0.05,  0.05)
+        glVertex3f(-0.025,  0.05,  0.05)
+
+        ## back
+        glColor3ub(self.red, self.green, self.blue)
+        glVertex3f( 0.025, -0.05, -0.05)
+        glVertex3f( 0.025, -0.05,  0.05)
+        glVertex3f( 0.025,  0.05,  0.05)
+        glVertex3f( 0.025,  0.05, -0.05)
+
+        ## front
+        glColor3ub(self.red, self.green, self.blue)
+        glVertex3f(-0.025, -0.05, -0.05)
+        glVertex3f(-0.025, -0.05,  0.05)
+        glVertex3f(-0.025,  0.05,  0.05)
+        glVertex3f(-0.025,  0.05, -0.05)
+        glEnd()
+
+        glPushMatrix()
+        glColor3ub(self.red, self.green, self.blue)
+        glTranslated(-0.05, 0.0, 0.0)
+        quadric = gluNewQuadric()
+        glRotatef(90.0, 0.0, 1.0, 0.0)
+        gluCylinder(quadric, 0.025, 0.025, 0.03, 16, 16)
+        gluDeleteQuadric(quadric)
+        glPopMatrix()
+
+        glPopMatrix()
+        
+    def getRotationAngle(self, v1, v2):
+        v1_u = self.getUnitVector(v1)
+        v2_u = self.getUnitVector(v2)
+        return np.degrees(np.arccos(np.clip(np.dot(v1_u, v2_u), -1, 1)))
+
+    def getUnitVector(self, vector):
+        return vector / np.linalg.norm(vector)
+
+    def onFocusCenter(self):
+        self.mode = 'rotate'
+        cameraCenterPoint = (self.x, self.y, self.z)
+        currentFacingPoint = (self.x - 0.5, self.y, self.z)
+        desirableFacingPoint = (0.0, 0.0, 0.0)
+
+        v1 = np.subtract(currentFacingPoint, cameraCenterPoint)
+        v2 = np.subtract(desirableFacingPoint, cameraCenterPoint)
+        
+        self.angle = self.getRotationAngle(v1, v2)
+        self.rotationVector = np.cross(self.getUnitVector(v1), self.getUnitVector(v2))
+        self.onDraw()
+
+    def getZbyAngle(self, angle):
+        return np.sqrt(np.square(0.5 / angle) - 0.25)
+
+    def onMove(self, axis, amount):
+        if axis in Axis and amount != 0:
+            if axis == Axis.X:
+                self.x += amount
+            elif axis == Axis.Y:
+                self.y += amount
+            elif axis == Axis.Z:
+                self.z += amount
+            elif axis == Axis.B:
+                self.b += amount
+            elif axis == Axis.C:
+                self.c += amount
