+++ canvas.py
@@ -1,0 +1,125 @@
+from OpenGL.GL import *
+from OpenGL.GLUT import *
+import wx
+from wx import glcanvas
+import numpy as np
+from camera import Camera
+
+class CanvasBase(glcanvas.GLCanvas):
+    def __init__(self, parent):
+        glcanvas.GLCanvas.__init__(self, parent, -1)
+        self.init = False
+        self.context = glcanvas.GLContext(self)
+        
+        # initial mouse position
+        self.lastx = self.x = 30
+        self.lasty = self.y = 30
+        self.size = None
+        self.Bind(wx.EVT_ERASE_BACKGROUND, self.OnEraseBackground)
+        self.Bind(wx.EVT_SIZE, self.OnSize)
+        self.Bind(wx.EVT_PAINT, self.OnPaint)
+        self.Bind(wx.EVT_LEFT_DOWN, self.OnMouseDown)
+        self.Bind(wx.EVT_LEFT_UP, self.OnMouseUp)
+        self.Bind(wx.EVT_MOTION, self.OnMouseMotion)
+
+    def OnEraseBackground(self, event):
+        pass # Do nothing, to avoid flashing on MSW.
+
+    def OnSize(self, event):
+        wx.CallAfter(self.DoSetViewport)
+        event.Skip()
+
+    def DoSetViewport(self):
+        size = self.size = self.GetClientSize()
+        self.SetCurrent(self.context)
+        glViewport(0, 0, size.width, size.height)
+        
+    def OnPaint(self, event):
+        self.SetCurrent(self.context)
+        if not self.init:
+            self.InitGL()
+            self.init = True
+        self.OnDraw()
+
+    def OnMouseDown(self, evt):
+        self.CaptureMouse()
+        self.x, self.y = self.lastx, self.lasty = evt.GetPosition()
+
+    def OnMouseUp(self, evt):
+        self.ReleaseMouse()
+
+    def OnMouseMotion(self, evt):
+        if evt.Dragging() and evt.LeftIsDown():
+            self.lastx, self.lasty = self.x, self.y
+            self.x, self.y = evt.GetPosition()
+            self.Refresh(False)
+
+
+class Canvas(CanvasBase):
+    def InitGL(self):
+        # set viewing projection
+        glMatrixMode(GL_PROJECTION)
+        glFrustum(-0.5, 0.5, -0.5, 0.5, 0.4, 5)
+
+        # position viewer
+        glMatrixMode(GL_MODELVIEW)
+        glTranslatef(0.0, 0.0, -1.5)
+
+        # position object
+        glRotatef(self.y, 1.0, 0.0, 0.0)
+        glRotatef(self.x, 0.0, 1.0, 0.0)
+
+        self.cameras = []
+
+    def OnDraw(self):
+        # clear color and depth buffers
+        glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT)
+
+        self.InitGrid()
+
+        if self.size is None:
+            self.size = self.GetClientSize()
+        w, h = self.size
+        w = max(w, 1.0)
+        h = max(h, 1.0)
+        xScale = 180.0 / w
+        yScale = 180.0 / h
+
+        ## object
+        glPushMatrix()
+        glColor3ub(0, 0, 128)
+        glTranslated(0.0,0.0,0.0)
+        glutSolidSphere(0.2,60,60)
+        glPopMatrix()
+        glRotatef((self.y - self.lasty) * yScale, 1.0, 0.0, 0.0)
+        glRotatef((self.x - self.lastx) * xScale, 0.0, 1.0, 0.0)
+
+        for camera in self.cameras:
+            camera.onDraw()
+
+        self.SwapBuffers()
+
+    def InitGrid(self):
+        glColor3ub(255, 255, 255)
+        
+        glBegin(GL_LINES)
+        for i in np.arange(-1, 1, 0.05):
+            glVertex3f(i, 0, 1)
+            glVertex3f(i, 0, -1)
+            glVertex3f(1, 0, i)
+            glVertex3f(-1, 0, i)
+        glColor3ub(255, 0, 0)
+        glVertex3f(0, 0, 0)
+        glVertex3f(1.5, 0, 0)
+        glColor3ub(0, 255, 0)
+        glVertex3f(0, 0, 0)
+        glVertex3f(0, 1.5, 0)
+        glColor3ub(0, 0, 255)
+        glVertex3f(0, 0, 0)
+        glVertex3f(0, 0, 1.5)
+        glEnd()
+
+    def OnDrawCamera(self, x, y, z, b, c):
+        camera = Camera(x, y, z, b, c)
+        self.cameras.append(camera)
+        self.OnDraw()
