+++ enums.py
@@ -1,0 +1,10 @@
+from enum import Enum
+
+class Axis(Enum):
+    X = "x"
+    Y = "y"
+    Z = "z"
+    B = "b"
+    C = "c"
+    Plus = "++"
+    Minus = "-"
